requirements.txt
streamlit==1.28.0
openai==0.28.0
faiss-cpu==1.7.4
numpy==1.24.0
python-dotenv==1.0.0
langchain==0.0.200
PyPDF2==3.0.0
tqdm==4.65.0streamlit==1.28.0
openai==0.28.0
faiss-cpu==1.7.4
numpy==1.24.0
python-dotenv==1.0.0
langchain==0.0.200
PyPDF2==3.0.0
tqdm==4.65.0
requests==2.28.0
beautifulsoup4==4.11.0